process ptsd data
